Move test utils for the past into dedicated `Past` class
