+++ client/src/test/java/org/spine3/change/ChangesShould.java
@@ -25,6 +25,7 @@ import com.google.protobuf.Timestamp;
 import org.junit.Test;
 import org.spine3.protobuf.Timestamps2;
 import org.spine3.test.NullToleranceTest;
+import org.spine3.test.TimeTests;
 
 import java.util.UUID;
 
@@ -122,7 +123,7 @@ public class ChangesShould {
 
     @Test
     public void create_TimestampChange_instance() {
-        final Timestamp fiveMinutesAgo = Timestamps2.minutesAgo(5);
+        final Timestamp fiveMinutesAgo = TimeTests.Past.minutesAgo(5);
         final Timestamp now = Timestamps2.getCurrentTime();
 
         final TimestampChange result = Changes.of(fiveMinutesAgo, now);
