+++ server/src/test/java/org/spine3/server/command/CommandSchedulingShould.java
@@ -45,10 +45,10 @@ import static org.spine3.base.CommandStatus.SCHEDULED;
 import static org.spine3.base.Commands.setSchedule;
 import static org.spine3.base.Identifiers.newUuid;
 import static org.spine3.protobuf.Durations2.minutes;
-import static org.spine3.protobuf.Timestamps2.minutesAgo;
 import static org.spine3.server.command.Given.Command.addTask;
 import static org.spine3.server.command.Given.Command.createProject;
 import static org.spine3.server.command.Given.Command.startProject;
+import static org.spine3.test.TimeTests.Past.minutesAgo;
 
 public class CommandSchedulingShould extends AbstractCommandBusTestSuite {
