+++ testutil/src/test/java/org/spine3/server/aggregate/AggregateBuilderShould.java
@@ -23,6 +23,7 @@ package org.spine3.server.aggregate;
 import com.google.protobuf.Timestamp;
 import org.junit.Test;
 import org.spine3.protobuf.Timestamps2;
+import org.spine3.test.TimeTests;
 
 import static org.junit.Assert.assertEquals;
 
@@ -39,7 +40,7 @@ public class AggregateBuilderShould {
         final int id = 2048;
         final int version = 2017;
         final Timestamp whenModified = Timestamps2.getCurrentTime();
-        final Timestamp state = Timestamps2.minutesAgo(60);
+        final Timestamp state = TimeTests.Past.minutesAgo(60);
 
         final Aggregate aggregate = givenAggregate()
                 .withId(id)
